requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Babel==2.11.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.16
django-allauth==0.51.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-environ==0.9.0
django-extra-views==0.13.0
django-filter==22.1
django-haystack==3.2.1
django-js-asset==2.0.0
django-oscar==3.1
django-phonenumber-field==5.2.0
django-solo==2.0.0
django-tables2==2.3.4
django-treebeard==4.4
django-utils-six==2.0
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==15.3.3
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
phonenumbers==8.13.0
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg==3.1.4
psycopg-binary==3.1.4
psycopg2-binary==2.9.5
purl==1.6
pycountry==22.3.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
stripe==5.0.0
typing-extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
